requirements.txt
langchain
langchain_core
langchain_openai
openai
langchain-anthropic
langchain-google-genai
google-generativeai
langchain_huggingface
transformers
huggingface-hub
python_dotenv
streamlit
scikit-learn
numpy
langchain-ollama